instafan_python/requirements.txt
uvicorn
email-validator
fastapi
PyJWT
python-decouple
mysql-connector-python
pycryptodome
bcrypt
python-multipart
google-cloud-storage